% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_capacc.R
\name{capa.cc}
\alias{capa.cc}
\title{Collective and point anomalies in cross-correlated data---CAPA-CC}
\usage{
capa.cc(
  x,
  Q,
  b = 1,
  b_point = 1,
  min_seg_len = 2,
  max_seg_len = 10^8,
  transform = centralise
)
}
\arguments{
\item{x}{An n x p data matrix where each row is an observation vector.}

\item{Q}{An estimate of the precision matrix. See \code{\link{robust_sparse_precision}}. Must be a sparse matrix from the Matrix package.}

\item{b}{The scaling factor for the collective anomaly penalty. Defaults to 1.}

\item{b_point}{The scaling factor for the point anomaly penalty. Defaults to 1.}

\item{min_seg_len}{The minimum segment length. Defaults to 2.}

\item{max_seg_len}{The maximum segment length. Defaults to 10^8.}

\item{transform}{A function used to centre the data prior to analysis by \code{capa.cc}. This can, for example, be used to compensate for the effects of autocorrelation in the data.
The default value is \code{transform=centralise}, which centrailises the data by the median.
Other choices are available in the anomaly package, for example \code{anomaly::robustscale} and \code{anomaly::ac_corrected}), and a user defined function can also be specified.}
}
\value{
An S3 class of type capacc with the following components:
\describe{
\item{\code{x}}{The input data matrix.}
\item{\code{anoms}}{A data frame with four columns: start (start-point of the anomaly), end (end-point of an anomaly), variate (which variable is affected) and size (the estimated size of the mean component for the given variate).}
}
}
\description{
A method for detecting anomalous segments and points based on CAPA-CC by Tveten, Eckley, Fearnhead (2020).
}
\examples{
library(capacc)
x <- simulate_cor()$x
Q <- robust_sparse_precision(x, adjacency_mat(banded_neighbours(2, ncol(x)), sparse = FALSE))
res <- capa.cc(x, Q, b = 1, min_seg_len = 5)
plot(res)
collective_anomalies(res)

}
\references{

}
