% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{capacc}
\alias{capacc}
\title{CAPA-CC}
\usage{
capacc(x, Q, b = 1, b_point = 1, min_seg_len = 2L, max_seg_len = 100000000L)
}
\arguments{
\item{x}{An n x p data matrix where each row is an observation vector.}

\item{Q}{An estimate of the precision matrix. See \link{\code{estimate_precision_mat}}}

\item{b}{The scaling factor for the collective anomaly penalty. Defaults to 1.}

\item{b_point}{The scaling factor for the point anomaly penalty. Defaults to 1.}

\item{min_seg_len}{The minimum segment length. Defaults to 2.}

\item{max_seg_len}{The maximum segment length. Defaults to 10^8.}
}
\value{
A data frame containg start- and end-points of anaomlies, indices of which variables are affected as well as the size of the anomalous mean for each variable.
}
\description{
CAPA-CC
}
