% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mat_generators.R
\name{car_precision_mat}
\alias{car_precision_mat}
\title{Generate a row-standardised CAR precision matrix}
\usage{
car_precision_mat(
  nbs = banded_neighbours(2, 10),
  rho = 0.5,
  sigma = 1,
  standardised = TRUE,
  sparse = TRUE
)
}
\arguments{
\item{nbs}{A list where nbs[[i]] is a vector with the neighbours of variable i. The functions \code{banded_neighbours}, \code{lattice_neighbours}, \code{random_neighbours} can be used to generate such lists of neighbours for banded, lattice and random adjacency structures.}

\item{rho}{The parameter in the row-standardised CAR model.}

\item{sigma}{The standard deviation of each variables. Defaults to 1.}

\item{standardised}{A logical value for whether the output precision matrix should be stadnardised so that its inverse is a correlation matrix.}

\item{sparse}{A logical value specifying whether the output should be a sparse matrix from the Matrix package (TRUE) or not (FALSE).}
}
\value{
A CAR precision matrix.
}
\description{
Generate a row-standardised CAR precision matrix
}
