% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mat_generators.R
\name{cuthill_mckee}
\alias{cuthill_mckee}
\title{Cuthill McKee (CM) algorithm}
\usage{
cuthill_mckee(x, reverse = FALSE, return_all = FALSE)
}
\arguments{
\item{x}{Input matrix. Must be square and symmetric.}
}
\value{
A list with the reordered matrix 'x' and the ordering 'order'.
}
\description{
Transform sparse matrix into a banded matrix. Code inspired by the netprioR package.
}
\author{
Fabian Schmich
}
