% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_distr.R
\name{many_cpt_distr}
\alias{many_cpt_distr}
\title{#' @export}
\usage{
many_cpt_distr(
  out_file = "cpt_distr.csv",
  variables = list(cost = c("inspect", "inspect_iid", "mvlrt", "mvlrt_iid"), rho =
    c(0.7, 0.9, 0.99), proportions = c(0.1, 0.3, 1)),
  data = init_data(n = 100, locations = 50, durations = 50),
  method = method_params(),
  loc_tol = 10
)
}
\description{
#' @export
}
