% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpt_est.R
\name{many_cpt_est}
\alias{many_cpt_est}
\title{#' @export}
\usage{
many_cpt_est(
  out_file,
  variables,
  data = init_data(),
  method = method_params(),
  tuning = tuning_params(),
  n_sim = 100
)
}
\description{
#' @export
}
