% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.capacc}
\alias{plot.capacc}
\title{Plotting function for capacc objects}
\usage{
\method{plot}{capacc}(object, epoch = n, subset = 1:p, variate_names = NULL, true_anoms = NULL)
}
\arguments{
\item{object}{A capacc object to be plotted.}

\item{epoch}{Plots data from 1:epoch.}

\item{subset}{Plots variables in subset.}

\item{variate_names}{An optional vector with the names of the variables.}

\item{true_anoms}{An optional data frame with columns "start" and "end" to illustrate known locations of anomalies.}
}
\value{
A ggplot of the detected collective and point anomalies.
}
\description{
Plotting function for capacc objects
}
