% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect.R
\name{power.method}
\alias{power.method}
\title{Power method for finding the leading eigenvector of a symmetric matrix}
\usage{
power.method(A, eps = 1e-10, maxiter = 10000)
}
\arguments{
\item{A}{a square symmetric matrix}

\item{eps}{tolerance for convergence (in Frobenius norm)}

\item{maxiter}{maximum iteration}
}
\value{
a unit-length leading eigenvector of A
}
\description{
Power method for finding the leading eigenvector of a symmetric matrix
}
