% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cor_mat_estimators.R
\name{robust_sparse_precision}
\alias{robust_sparse_precision}
\title{Robustly estimate a precision matrix with a given adjacency structure.}
\usage{
robust_sparse_precision(x, adj_mat, sparse = TRUE)
}
\arguments{
\item{x}{An n x p data matrix where each row is an observation vector.}

\item{adj_mat}{A p x p adjacency matrix. See details.}

\item{sparse}{A logical specifying whether the returned estimate should be a sparse matrix from the Matrix package or not.}
}
\value{
A precision matrix.
}
\description{
The function uses the GLASSAO algorithm implemented in the glassoFast package to estimate a precision matrix robustly with a conditional independence structure specified by \code{adj_mat}.
}
\details{
The function \code{adjacency_mat} can be used to generate an adjacency matrix from a list containing the neighbours of variable i in list[[i]].
Moreover, the functions \code{banded_neighbours}, \code{lattice_neighbours}, \code{random_neighbours} can be used to generate such lists of neighbours for banded, lattice and random adjacency structures.
}
\examples{
x <- simulate_cor()$x
Q <- robust_sparse_precision(x, adjacency_mat(banded_neighbours(2, ncol(x)), sparse = FALSE))

}
