% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_inspect.R
\name{vector.norm}
\alias{vector.norm}
\title{Norm of a vector}
\usage{
vector.norm(v, q = 2, na.rm = FALSE)
}
\arguments{
\item{v}{a vector of real numbers}

\item{q}{a nonnegative real number or Inf}

\item{na.rm}{boolean, whether to remove NA before calculation}
}
\value{
the entrywise L_q norm of a vector or a matrix
}
\description{
Calculate the entrywise L_q norm of a vector or a matrix
}
