% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/method_inspect.R
\name{vector.soft.thresh}
\alias{vector.soft.thresh}
\title{Soft thresholding a vector}
\usage{
vector.soft.thresh(x, lambda)
}
\arguments{
\item{x}{a vector of real numbers}

\item{lambda}{soft thresholding value}
}
\value{
a vector of the same length
}
\description{
entries of v are moved towards 0 by the amount lambda until they hit 0.
}
